requirements.txt
alembic==1.13.3
fastapi==0.115.4
gunicorn==23.0.0
JSON-log-formatter==1.1
SQLAlchemy==2.0.36
uvicorn==0.32.0
python-dotenv==1.1.0
requests==2.28.2
openai==0.27.8